pyproject.toml
[project]
name = "blogsapp"
version = "0.1.0"
description = ""
authors = [
    {name = "1aerostorm",email = "aerostorm1@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "django (>=5.2.8,<6.0.0)",
    "djangorestframework (>=3.16.1,<4.0.0)",
    "mssql-django (>=1.6,<2.0)"
]

[tool.poetry]
packages = [{include = "blogsapp", from = "src"}]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
